Rename PeopleMover.tsx to SpacePage.tsx and PeopleMoverHeader.tsx to SpacePageHeader.tsx
