+++ ui/src/SpacePage/ProductList/ProductListGrouped/ProductListGrouped.scss
